requirements.txt
langchain>=1.2.15
langchain-core>=1.2.28
langgraph>=1.1.6
langchain-google-genai>=4.2.1
langchain-anthropic>=0.3.22
langchain-openai>=0.3.35
neo4j
pyDatalog
ipython
requests
pandas
python-dotenv
langchain-text-splitters>=0.3.11
langgraph-checkpoint>=2.1.2
langgraph-prebuilt>=0.6.5
langgraph-sdk>=0.3.13
langsmith>=0.4.37
google-genai>=1.47.0
ipykernel
